Move "checkForEGL()" below open command handler
Also ignore help command mode.

It prevents waking up dGPU from suspended state when we want to display
help message or we want to use IPC to open a file in existing QMPlay2
instance.

It's called after creating the QApplication instance, but it is called
before Qt reads specified envvars.